+++ packages/core/loglayer/src/index.ts
@@ -21,6 +21,7 @@ export { LogLayer } from "./LogLayer.js";
 export { MockLogBuilder } from "./MockLogBuilder.js";
 export { MockLogLayer } from "./MockLogLayer.js";
 export { TestLoggingLibrary } from "./TestLoggingLibrary.js";
+export { BlankTransport } from "./transports/BlankTransport.js";
 export { ConsoleTransport } from "./transports/ConsoleTransport.js";
 export { TestTransport } from "./transports/TestTransport.js";
 export * from "./types/index.js";